Move the compatibility stuff into a new directory, or it'll get huge.
