+++ tsconfig.json
@@ -17,6 +17,7 @@
                 "src/components/*",
                 "src/views/*",
                 "src/data/*",
+                "src/compatibility/*",
                 "src/*"
             ],
             "requests": [
